+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## [0.0.8] - 2026-03-08
+
+### Added
+
+- Support for Posthog session recording
+- Support for base URL configuration
+- Support for HTTP Basic Auth configuration
+- Added the example folders
+
 ## [0.0.1] - 2026-03-08
 
 ### Added